application/requirements.txt
redis
Flask
cassandra-driver
